+++ R/GADSdat_utilities.R
@@ -34,11 +34,16 @@ check_GADSdat <- function(GADSdat) {
   }
 
   var_info <- unique(GADSdat$labels[, c("varName", "varLabel", "format", "display_width", "labeled")])
-  if(!nrow(var_info) == length(unique(names(GADSdat$dat)))) {
+  if(nrow(var_info) != length(unique(names(GADSdat$dat)))) {
     by(GADSdat$labels, GADSdat$labels$varName, function(labels) {
       if(nrow(unique(labels[, c("varName", "varLabel", "format", "display_width", "labeled")])) > 1) {
         stop("The following variable has inconsistent meta information on variable level: ", unique(labels$varName))
       }
     })
   }
+  by(GADSdat$labels, GADSdat$labels$varName, function(labels) {
+    if(any(duplicated(labels$value))) {
+      stop("The following variable has duplicate values rows in its meta data: ", unique(labels$varName))
+    }
+  })
 }